+++ firestore-pdf-generator/functions/lib/utilities/event_channel.ts
@@ -5,6 +5,6 @@ export const EVENT_TYPE_PREFIX =
 
 export const eventChannel = process.env.EVENTARC_CHANNEL
   ? getEventarc().channel(process.env.EVENTARC_CHANNEL, {
-    allowedEventTypes: process.env.EXT_SELECTED_EVENTS,
-  })
+      allowedEventTypes: process.env.EXT_SELECTED_EVENTS,
+    })
   : null;