+++ firestore-pdf-generator/functions/src/index.ts
@@ -3,9 +3,8 @@ import * as admin from "firebase-admin";
 
 admin.initializeApp();
 
-import * as fs from "fs";
 import * as functions from "firebase-functions";
-import { parseParameters } from "./parse_parameters";
+import { FirestoreDocument, parseParameters } from "./parse_parameters";
 import { producePdf } from "lib/produce_pdf";
 import { runAction } from "lib/utilities/action";
 import { createErrorHandler } from "lib/utilities/error_handler";
@@ -15,24 +14,6 @@ process.on("unhandledRejection", (reason, p) => {
   console.error(reason, "Unhandled Rejection at Promise", p);
 });
 
-import { resolve } from "path";
-
-/**
- * Hello
- * @param {string} dir
- */
-async function* getFiles(dir: string): AsyncGenerator<string> {
-  const dirents = fs.readdirSync(dir, { withFileTypes: true });
-  for (const dirent of dirents) {
-    const res = resolve(dir, dirent.name);
-    if (dirent.isDirectory()) {
-      yield* getFiles(res);
-    } else {
-      yield res;
-    }
-  }
-}
-
 exports.executePdfGenerator = functions.firestore
   .document(extensionParameters.FIRESTORE_COLLECTION)
   .onCreate(
@@ -40,26 +21,12 @@ exports.executePdfGenerator = functions.firestore
       snapshot: functions.firestore.QueryDocumentSnapshot,
       context: functions.EventContext
     ) => {
-      console.log(__dirname);
-      (async () => {
-        for await (const f of getFiles(__dirname)) {
-          console.log(f);
-        }
-      })();
-
-      console.log(123, ".");
-      (async () => {
-        for await (const f of getFiles(".")) {
-          console.log(f);
-        }
-      })();
-
       const id = snapshot.id;
-      const data = snapshot.data();
+      const rawParameters = snapshot.data() as FirestoreDocument;
       const errorHandler = createErrorHandler({
         context: {
           id,
-          data,
+          rawParameters,
           context,
         },
       });
@@ -68,7 +35,8 @@ exports.executePdfGenerator = functions.firestore
         process.on("uncaughtException", errorHandler);
 
         const parameters = runAction(parseParameters, {
-          rawParameters: { data, id },
+          rawParameters,
+          id,
         });
 
         await runAction(producePdf, {