+++ scripts/update-documents.js
@@ -1,5 +1,6 @@
 #!/usr/bin/env node
 
+const YAML = require("yaml");
 const fs = require("fs");
 const extensions = require("./lib").extensions;
 
@@ -18,4 +19,34 @@ for (const extension of extensions) {
       (_, header, content) => header.trim() + content
     );
   fs.writeFileSync(`${extension}/CHANGELOG.md`, changelog);
+
+  const extensionFileContent = YAML.parse(
+    fs.readFileSync(`${extension}/extension.yaml`, { encoding: "utf8" })
+  );
+
+  const firebaseExtensionParametersHeader =
+    "## Firebase extension parameters\n";
+  let parametersMarkdown = firebaseExtensionParametersHeader;
+  for (const parameter of extensionFileContent["params"]) {
+    parametersMarkdown += `
+### \`${parameter.param}\` ${parameter.required ? "(required)" : "(optional)"}
+
+${parameter.label}<br/>
+type: **${parameter.type}**
+
+${parameter.description}
+`;
+  }
+
+  const preinstallTemplate = fs.readFileSync(`${extension}/PARAMETERS.md`, {
+    encoding: "utf8",
+  });
+  const preinstallContent = preinstallTemplate.replace(
+    new RegExp(
+      `^${firebaseExtensionParametersHeader}.*?(.(?=\n#{1,3} )|$(?![\r\n]))`,
+      "ms"
+    ),
+    parametersMarkdown
+  );
+  fs.writeFileSync(`${extension}/PARAMETERS.md`, preinstallContent);
 }